user authentication backend

+++ controllers/expense.js
@@ -1,7 +1,7 @@
 const ExpenseItems=require('../models/expense');
 
 exports.getExpenses=(req,res,next)=>{
-    ExpenseItems.findAll().then(result=>{
+    ExpenseItems.findAll({where:{userId:req.user.id}}).then(result=>{
       res.status(201).json({newExpenseDetail:result})
       console.log(result)
     })
@@ -12,7 +12,8 @@ exports.postExpenses=(req,res,next)=>{
     ExpenseItems.create({
         expense:req.body.expense,
         description:req.body.description,
-        category:req.body.category
+        category:req.body.category,
+        userId:req.user.id
     }).then(result=>{
         res.status(201).json({newExpenseDetail:result})
         console.log(result)
@@ -22,7 +23,7 @@ exports.postExpenses=(req,res,next)=>{
 
 exports.getDelete=(req,res,next)=>{
     const userId=req.params.userId;
-    ExpenseItems.destroy({where:{id:userId}}).then((result)=>{
+    ExpenseItems.destroy({where:{id:userId,userId:req.user.id}}).then((result)=>{
        console.log(result)
        console.log('deleted')
     })
